Rename Add Team Member to Mensen Verzamelen and its texts
